Form1.Designer: make the form to auto resize
Resize the form[0] according to the setting of AutoSizeMode, which in our
case is GrowOnly.

The control grows as much as necessary to fit its contents but does not
shrink smaller than the value of its Size property. The form can be resized,
but cannot be made so small that any of its contained controls are hidden.[1]

[0]: https://docs.microsoft.com/en-us/dotnet/api/system.windows.forms.form.autosize?view=windowsdesktop-5.0
[1]: https://docs.microsoft.com/en-us/dotnet/api/system.windows.forms.autosizemode?view=windowsdesktop-5.0